Refactor: Move TradeValidation to trade package
