+++ core/src/main/java/bisq/core/trade/TradeManager.java
@@ -132,7 +132,7 @@ import org.jetbrains.annotations.NotNull;
 
 import javax.annotation.Nullable;
 
-import static bisq.core.trade.protocol.bisq_v1.TradeValidation.checkInputsForDepositTxRequest;
+import static bisq.core.trade.TradeValidation.checkInputsForDepositTxRequest;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
