+++ core/src/main/java/bisq/core/trade/model/bisq_v1/Trade.java
@@ -76,7 +76,7 @@ import org.jetbrains.annotations.NotNull;
 
 import javax.annotation.Nullable;
 
-import static bisq.core.trade.protocol.bisq_v1.TradeValidation.checkTxFee;
+import static bisq.core.trade.TradeValidation.checkTxFee;
 import static com.google.common.base.Preconditions.checkNotNull;
 
 /**
