+++ core/src/main/java/bisq/core/trade/protocol/bisq_v1/DisputeProtocol.java
@@ -49,7 +49,7 @@ import bisq.common.handlers.ResultHandler;
 
 import lombok.extern.slf4j.Slf4j;
 
-import static bisq.core.trade.protocol.bisq_v1.TradeValidation.checkTradeId;
+import static bisq.core.trade.TradeValidation.checkTradeId;
 
 @Slf4j
 public class DisputeProtocol extends TradeProtocol {
