+++ core/src/main/java/bisq/core/trade/protocol/bisq_v1/tasks/maker/MakerProcessesInputsForDepositTxRequest.java
@@ -44,7 +44,7 @@ import java.util.Optional;
 
 import lombok.extern.slf4j.Slf4j;
 
-import static bisq.core.trade.protocol.bisq_v1.TradeValidation.*;
+import static bisq.core.trade.TradeValidation.*;
 import static com.google.common.base.Preconditions.checkNotNull;
 
 @Slf4j
