+++ core/src/main/java/bisq/core/trade/protocol/bisq_v1/tasks/seller/SellerProcessCounterCurrencyTransferStartedMessage.java
@@ -27,7 +27,7 @@ import bisq.common.taskrunner.TaskRunner;
 
 import lombok.extern.slf4j.Slf4j;
 
-import static bisq.core.trade.protocol.bisq_v1.TradeValidation.checkBitcoinAddress;
+import static bisq.core.trade.TradeValidation.checkBitcoinAddress;
 import static com.google.common.base.Preconditions.checkNotNull;
 
 @Slf4j
